Renamed `coi.js` to `crossOriginIsolation.js`
